refactor: restructure settings components and enhance UI for sectors and groups
- Updated the routing to point to the new Settings component structure.
- Refactored the Settings view to improve layout and user experience, including the introduction of a new tabbed interface for managing sectors and groups.
- Removed deprecated components related to group and sector management, streamlining the codebase.
- Enhanced the UI with new components for better organization and accessibility of settings options.
- Improved localization and overall maintainability of the settings components.
