+++ src/views/Settings/CustomBreaks/DeleteCustomBreakModal.vue
