+++ src/views/Settings/Sectors/index.vue
