--- src/views/SettingsV2/SettingsProjectOptions/SettingsProjectOptionsItem.vue
@@ -1,54 +1,0 @@
-<template>
-  <section class="project-options__item">
-    <UnnnicSwitch
-      :modelValue="props.modelValue"
-      :textRight="props.name"
-      :disabled="props.disabled"
-      @update:model-value="emit('update:model-value', $event)"
-    />
-    <UnnnicToolTip
-      v-if="props.tooltip"
-      enabled
-      :text="props.tooltip"
-      side="right"
-      maxWidth="20rem"
-    >
-      <UnnnicIcon
-        icon="info"
-        scheme="fg-base"
-        filled
-        size="sm"
-      />
-    </UnnnicToolTip>
-  </section>
-</template>
-
-<script setup lang="ts">
-interface Props {
-  modelValue: boolean;
-  name: string;
-  tooltip?: string;
-  disabled?: boolean;
-}
-
-const props = withDefaults(defineProps<Props>(), {
-  disabled: false,
-  tooltip: '',
-});
-
-const emit = defineEmits<{
-  'update:model-value': [boolean];
-}>();
-</script>
-
-<style lang="scss" scoped>
-.project-options__item {
-  display: flex;
-  gap: $unnnic-space-1;
-  align-items: center;
-
-  .unnnic-tooltip {
-    display: flex;
-  }
-}
-</style>
